+++ models/permissionModel.js
@@ -8,6 +8,7 @@ export const getAllPermissions = async () => {
 };
 
 export const getPermissionById = async (id) => {
+  if (!id) throw new Error('Permission ID is required');
   const [rows] = await db.query(
     'SELECT * FROM permissions WHERE permission_id = ?',
     [id]
@@ -17,6 +18,9 @@ export const getPermissionById = async (id) => {
 
 export const createPermission = async (data) => {
   const { permission_name, description } = data;
+  
+  if (!permission_name) throw new Error('Permission name is required');
+  
   const [result] = await db.query(
     `INSERT INTO permissions (permission_name, description)
      VALUES (?, ?)`,
@@ -26,6 +30,8 @@ export const createPermission = async (data) => {
 };
 
 export const updatePermission = async (id, data) => {
+  if (!id) throw new Error('Permission ID is required');
+  
   const { permission_name, description } = data;
   const [result] = await db.query(
     `UPDATE permissions
@@ -37,6 +43,7 @@ export const updatePermission = async (id, data) => {
 };
 
 export const deletePermission = async (id) => {
+  if (!id) throw new Error('Permission ID is required');
   const [result] = await db.query(
     'DELETE FROM permissions WHERE permission_id = ?',
     [id]
